scripts for setting gpio permissions
Here's a udev hook that allows writing to the gpio pins for normal
users.

There's a catch to it, however: When the gpioN directory appears (so the
export call returns), the sysfs files still are owned by root, this is
only done async by udev (by these rules). The kgpio library takes care
of this by waiting until the file is writable, before using the pins.